requirements.txt
pillow==11.0.0
PyMuPDF==1.25.1
